pyproject.toml
[tool.poetry]
name = "azure-ad-monitor"
version = "0.1.0"
description = ""
authors = ["Sean Johnson <seanson@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.7"
azure-core = "1.2.2"
azure-graphrbac = "^0.61.1"
azure-cli-core = "^2.1.0"
pytz = "^2019.3"
Flask-APScheduler = "^1.11.0"
prometheus_client = "^0.7.1"
uwsgi = "^2.0.18"

[tool.poetry.dev-dependencies]
black = {version = "^19.10b0", allow-prereleases = true}
flake8 = "^3.7.9"
isort = "^4.3.21"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
